Adding guards to only run File#chown and File.chown specs as root.

+++ spec/ruby/1.8/core/file/chown_spec.rb
@@ -1,46 +1,51 @@
 require File.dirname(__FILE__) + '/../../spec_helper'
 
-def group_zero
-  # HACK this is probably not the nicest way to find the name of the group
-  # with id 0, but it works on OS X.
-  `cat /etc/group`.split(/\n|\r/).reject { |x| !(x =~ /:0:[^:]*$/) }.first.gsub(/^([^:]*):.*$/, '\1')
-end
+# Only run this spec as root, as it may fail for other users
+if `whoami` == "root\n"
 
-describe "File.chown" do
-  
-  before(:each) do
-    system "echo 'rubinius' > /tmp/chown_test"
-    system "chown #{`whoami`.gsub(/\n/, '')} /tmp/chown_test"
+  def group_zero
+    # HACK this is probably not the nicest way to find the name of the group
+    # with id 0, but it works on OS X.
+    `cat /etc/group`.split(/\n|\r/).reject { |x| !(x =~ /:0:[^:]*$/) }.first.gsub(/^([^:]*):.*$/, '\1')
   end
+
+  describe "File.chown" do
   
-  after(:each) do
-    system 'rm /tmp/chown_test' if File.exist?("/tmp/chown_test")
+    before(:each) do
+      system "echo 'rubinius' > /tmp/chown_test"
+      system "chown #{`whoami`.gsub(/\n/, '')} /tmp/chown_test"
+    end
+  
+    after(:each) do
+      system 'rm /tmp/chown_test' if File.exist?("/tmp/chown_test")
+    end
+  
+    it "should change the ownerid and groupid of the file" do
+      File.chown(nil, 0, '/tmp/chown_test').should == 1
+      `ls -l /tmp/chown_test`.should =~ %r{#{group_zero}}
+    end
+  
   end
-  
-  it "should change the ownerid and groupid of the file" do
-    File.chown(nil, 0, '/tmp/chown_test').should == 1
-    `ls -l /tmp/chown_test`.should =~ %r{#{group_zero}}
-  end
-  
-end
 
 
-describe "File#chown" do
+  describe "File#chown" do
   
-  before(:each) do
-    system "echo 'rubinius' > /tmp/chown_test"
-    system "chown #{`whoami`.gsub(/\n/, '')} /tmp/chown_test"
+    before(:each) do
+      system "echo 'rubinius' > /tmp/chown_test"
+      system "chown #{`whoami`.gsub(/\n/, '')} /tmp/chown_test"
+    end
+  
+    after(:each) do
+      system 'rm /tmp/chown_test' if File.exist?("/tmp/chown_test")
+    end
+  
+    it "should change the ownerid and groupid of the file" do
+      f = File.open('/tmp/chown_test', 'r')
+      f.chown(nil, 0).should == 0
+      f.close
+      `ls -l /tmp/chown_test`.should =~ %r{#{group_zero}}
+    end
+  
   end
-  
-  after(:each) do
-    system 'rm /tmp/chown_test' if File.exist?("/tmp/chown_test")
-  end
-  
-  it "should change the ownerid and groupid of the file" do
-    f = File.open('/tmp/chown_test', 'r')
-    f.chown(nil, 0).should == 0
-    f.close
-    `ls -l /tmp/chown_test`.should =~ %r{#{group_zero}}
-  end
-  
+
 end
